+++ src/mongodb/IdentifiableMongoDbPersistence.ts
@@ -1,5 +1,10 @@
+let _ = require('lodash');
 let async = require('async');
 
+import { ConfigParams } from 'pip-services-commons-node';
+import { FilterParams } from 'pip-services-commons-node';
+import { PagingParams } from 'pip-services-commons-node';
+import { DataPage } from 'pip-services-commons-node';
 import { IIdentifiable } from 'pip-services-commons-node';
 import { IStringIdentifiable } from 'pip-services-commons-node';
 import { ObjectWriter } from 'pip-services-commons-node';
@@ -12,14 +17,82 @@ import { IWriter, IGetter, ISetter } from '../.';
 
 export class IdentifiableMongoDbPersistence<T extends IIdentifiable<K>, K> extends MongoDbPersistence
     implements IWriter<T, K>, IGetter<T, K>, ISetter<T> {
+    protected _maxPageSize: number = 100;
 
     public constructor(collection: string, schema: Schema) {
+        super(collection, schema);
+
         if (collection == null)
             throw new Error("Collection name could not be null");
         if (schema == null)
             throw new Error("Schema could not be null");
+    }
 
-        super(collection, schema);
+    public configure(config: ConfigParams): void {
+        super.configure(config);
+        
+        this._maxPageSize = config.getAsIntegerWithDefault("max_page_size", this._maxPageSize);
+    }
+
+    protected getPageByFilter(correlationId: string, filter: any, paging: PagingParams, 
+        sort: any, select: any, callback: (err: any, items: DataPage<T>) => void): void {
+        // Adjust max item count based on configuration
+        paging = paging || new PagingParams();
+        let skip = paging.getSkip(-1);
+        let take = paging.getTake(this._maxPageSize);
+        let pagingEnabled = paging.total;
+
+        // Configure statement
+        let statement = this._model.find(filter);
+
+        if (skip >= 0) statement.skip(skip);
+        statement.limit(take);
+        if (sort && !_.isEmpty(sort)) statement.sort(sort);
+        if (select && !_.isEmpty(select)) statement.select(select);
+
+        statement.exec((err, items) => {
+            if (err) {
+                callback(err, null);
+                return;
+            }
+
+            items = _.map(items, this.convertToPublic);
+
+            if (pagingEnabled) {
+                this._model.count(filter, (err, count) => {
+                    if (err) {
+                        callback(err, null);
+                        return;
+                    }
+                        
+                    let page = new DataPage<T>(items, count);
+                    callback(null, page);
+                });
+            } else {
+                let page = new DataPage<T>(items);
+                callback(null, page);
+            }
+        });
+    }
+
+    protected getListByFilter(correlationId: string, filter: any, sort: any, select: any, 
+        callback: (err: any, items: T[]) => void): void {
+        
+        // Configure statement
+        let statement = this._model.find(filter);
+
+        if (sort && !_.isEmpty(sort)) statement.sort(sort);
+        if (select && !_.isEmpty(select)) statement.select(select);
+
+        statement.exec((err, items) => {
+            if (err) {
+                callback(err, null);
+                return;
+            }
+
+            items = _.map(items, this.convertToPublic);
+            callback(null, items);
+        });
     }
 
     public getOneById(correlationId: string, id: K, callback: (err: any, item: T) => void): void {
@@ -27,8 +100,29 @@ export class IdentifiableMongoDbPersistence<T extends IIdentifiable<K>, K> exten
             if (!err)
                 this._logger.trace(correlationId, "Retrieved from %s with id = %s", this._collection, id);
 
-            item = this.jsonToPublic(item);
+            item = this.convertToPublic(item);
             callback(err, item);
+        });
+    }
+
+    protected getOneRandom(correlationId: string, filter: any, callback: (err: any, item: T) => void): void {
+        this._model.count(filter, (err, count) => {
+            if (err) {
+                callback(err, null);
+                return;
+            }
+
+            let pos = _.random(0, count - 1);
+
+            this._model.find(filter)
+                .skip(pos)
+                .limit(1)
+                .exec((err, items) => {
+                    let item = (items != null && items.length > 0) ? items[0] : null;
+                    
+                    item = this.convertToPublic(item);
+                    callback(err, item);
+                });
         });
     }
 
@@ -47,7 +141,7 @@ export class IdentifiableMongoDbPersistence<T extends IIdentifiable<K>, K> exten
             if (!err)
                 this._logger.trace(correlationId, "Created in %s with id = %s", this._collection, newItem.id);
 
-            newItem = this.jsonToPublic(newItem);
+            newItem = this.convertToPublic(newItem);
             callback(err, newItem);
         });
     }
@@ -77,7 +171,7 @@ export class IdentifiableMongoDbPersistence<T extends IIdentifiable<K>, K> exten
                 this._logger.trace(correlationId, "Set in %s with id = %s", this._collection, item.id);
            
             if (callback) {
-                newItem = this.jsonToPublic(newItem);
+                newItem = this.convertToPublic(newItem);
                 callback(err, newItem);
             }
         });
@@ -98,7 +192,7 @@ export class IdentifiableMongoDbPersistence<T extends IIdentifiable<K>, K> exten
                 this._logger.trace(correlationId, "Update in %s with id = %s", this._collection, item.id);
 
             if (callback) {
-                newItem = this.jsonToPublic(newItem);
+                newItem = this.convertToPublic(newItem);
                 callback(err, newItem);
             }
         });
@@ -110,7 +204,7 @@ export class IdentifiableMongoDbPersistence<T extends IIdentifiable<K>, K> exten
                 this._logger.trace(correlationId, "Deleted from %s with id = %s", this._collection, id);
 
             if (callback) {
-                oldItem = this.jsonToPublic(oldItem);
+                oldItem = this.convertToPublic(oldItem);
                 callback(err, oldItem);
             }
         });